Change algorithm, add multiple key support
The algorithm now supports the OpenSSH algorithm, where 2-bit pairs are
read in little endian, rather than reading the whole string in big endian.
As a result, documentation is updated to reflect this.

The script now accepts multiple files, but passing -i|--id is required, as
future support for -k|--keyring will be needed as an alternative.